+++ src/providers/twitch/TwitchIrcServer.cpp
@@ -103,6 +103,22 @@ void sendHelixMessage(const std::shared_ptr<TwitchChannel> &channel,
         });
 }
 
+/// Returns true if chat messages should be sent over Helix
+bool shouldSendHelixChat()
+{
+    switch (getSettings()->chatSendProtocol)
+    {
+        case ChatSendProtocol::Helix:
+            return true;
+        case ChatSendProtocol::Default:
+        case ChatSendProtocol::IRC:
+            return false;
+        default:
+            assert(false && "Invalid chat protocol value");
+            return false;
+    }
+}
+
 }  // namespace
 
 namespace chatterino {
@@ -580,7 +596,7 @@ void TwitchIrcServer::onMessageSendRequested(
         return;
     }
 
-    if (getSettings()->enableHelixChatSend)
+    if (shouldSendHelixChat())
     {
         sendHelixMessage(channel, message);
     }
@@ -604,7 +620,7 @@ void TwitchIrcServer::onReplySendRequested(
         return;
     }
 
-    if (getSettings()->enableHelixChatSend)
+    if (shouldSendHelixChat())
     {
         sendHelixMessage(channel, message, replyId);
     }