+++ src/singletons/Settings.hpp
@@ -62,6 +62,12 @@ enum UsernameRightClickBehavior : int {
     Ignore = 2,
 };
 
+enum class ChatSendProtocol : int {
+    Default = 0,
+    IRC = 1,
+    Helix = 2,
+};
+
 /// Settings which are availlable for reading and writing on the gui thread.
 // These settings are still accessed concurrently in the code but it is bad practice.
 class Settings
@@ -524,7 +530,8 @@ public:
         HelixTimegateOverride::Timegate,
     };
 
-    BoolSetting enableHelixChatSend = {"/misc/helixChatSend", true};
+    EnumStringSetting<ChatSendProtocol> chatSendProtocol = {
+        "/misc/chatSendProtocol", ChatSendProtocol::Default};
 
     BoolSetting openLinksIncognito = {"/misc/openLinksIncognito", 0};